An ugly hack; temporary fix until a more elegant solution can be put in place.
Background: in the old code this was yet-another-global-variable; the now object-style model removes the global, but now the signal handler can't get to it... so we need a temporary global for the call. This almost certainly isn't extensible, and eventually a real solution is needed... but it gets the job done for the moment.